Create a new subcommand init
